refactor carrier signal and midi input into one
