+++ src/app/ui/run-vocoder-form.tsx
@@ -7,42 +7,7 @@ import MidiInput from "./midi-input";
 import ModulatorSignalInput from "./modulator-signal-input";
 import CarrierSignalInput from "./carrier-signal-input";
 import ShowWaveform from "./show-waveform";
-import SignalInput from "./signal-input";
-
-const defaultCarrierSignals = [
-  {
-    name: "Sine wave",
-    label: "Sine wave",
-    filename: "/sine_example.wav",
-    imagename: "sine_waveform.png",
-  },
-  {
-    name: "Square wave",
-    label: "Square wave",
-    filename: "/square_example.wav",
-    imagename: "square_waveform.png",
-  },
-  {
-    name: "Triangle wave",
-    label: "Triangle wave",
-    filename: "/triangle_example.wav",
-    imagename: "triangle_waveform.png",
-  },
-  {
-    name: "Sawtooth wave",
-    label: "Sawtooth wave",
-    filename: "/sawtooth_example.wav",
-    imagename: "sawtooth_waveform.png",
-  },
-];
-
-const defaultModulatorSignals = [
-  {
-    name: "Hello example",
-    label: "Hello example",
-    filename: "/hello_example.wav",
-  },
-];
+import { base64ToArrayBuffer } from "./utils";
 
 function SubmitButton() {
   const { pending } = useFormStatus();
@@ -58,15 +23,6 @@ function SubmitButton() {
   );
 }
 
-function base64ToArrayBuffer(base64) {
-  var binaryString = atob(base64);
-  var bytes = new Uint8Array(binaryString.length);
-  for (var i = 0; i < binaryString.length; i++) {
-    bytes[i] = binaryString.charCodeAt(i);
-  }
-  return bytes.buffer;
-}
-
 export default function RunVocoderForm() {
   async function runVocoderWithFile(prevState: any, formData: FormData) {
     // add modulator signal to formData
@@ -75,11 +31,11 @@ export default function RunVocoderForm() {
     const blob = await fetch(modulatorSignalPath).then((r) => r.blob());
     formData.set("modulator-signal", blob);
 
-    // add midi input to formData
-    const midiInputPath = formData.get("midi-input");
-    if (!midiInputPath) return;
-    const blob2 = await fetch(midiInputPath).then((r) => r.blob());
-    formData.set("midi-input", blob2);
+    // add carrier signal to formData
+    const carrierSignalPath = formData.get("carrier-signal");
+    if (!carrierSignalPath) return;
+    const blob2 = await fetch(carrierSignalPath).then((r) => r.blob());
+    formData.set("carrier-signal", blob2);
 
     return await runVocoder(prevState, formData);
   }
@@ -93,22 +49,9 @@ export default function RunVocoderForm() {
   return (
     <div>
       <form action={dispatch} className="">
-        <div className="flex gap-8">
-          <SignalInput
-            initialSignals={defaultCarrierSignals}
-            defaultSignalName="Sine wave"
-            formInputLabel="Carrier Signal"
-            signalType="carrier"
-          />
-          <SignalInput
-            initialSignals={defaultModulatorSignals}
-            defaultSignalName="Hello example"
-            formInputLabel="Modulator Signal"
-            signalType="modulator"
-          />
-        </div>
+        <ModulatorSignalInput />
         <div className="mt-4">
-          <MidiInput />
+          <CarrierSignalInput />
         </div>
         <div className="flex items-center mt-4 justify-between">
           <Switch name="show-steps" value="true">
